requirements.txt
streamlit>=1.45.1
plotly>=6.1.0
pandas>=2.2.3
numpy>=2.2.6
scikit-learn>=1.6.1
scipy>=1.15.3
joblib>=1.5.0
threadpoolctl>=3.6.0
requests>=2.31.0
python-dotenv>=1.0.0
geopy>=2.4.1
matplotlib>=3.8.0 